Do not remove users in LDAP when queries return an empty result
closes #34764

+++ model/storage-private/src/main/java/org/keycloak/storage/UserStorageManager.java
@@ -128,6 +128,7 @@ public class UserStorageManager extends AbstractStorageManager<UserStorageProvid
         if (model == null) {
             // remove linked user with unknown storage provider.
             logger.debugf("Removed user with federation link of unknown storage provider '%s'", user.getUsername());
+            deleteInvalidUserCache(realm, user);
             deleteInvalidUser(realm, user);
             return null;
         }
@@ -149,7 +150,10 @@ public class UserStorageManager extends AbstractStorageManager<UserStorageProvid
 
         UserModel validated = importedUserValidation.validate(realm, user);
         if (validated == null) {
-            deleteInvalidUser(realm, user);
+            deleteInvalidUserCache(realm, user);
+            if (model.isRemoveInvalidUsersEnabled()) {
+                deleteInvalidUser(realm, user);
+            }
             return null;
         } else {
             return validated;
@@ -204,14 +208,16 @@ public class UserStorageManager extends AbstractStorageManager<UserStorageProvid
         return result;
     }
 
-    protected void deleteInvalidUser(final RealmModel realm, final UserModel user) {
-        String userId = user.getId();
-        String userName = user.getUsername();
+    protected void deleteInvalidUserCache(final RealmModel realm, final UserModel user) {
         UserCache userCache = UserStorageUtil.userCache(session);
         if (userCache != null) {
             userCache.evict(realm, user);
         }
+    }
 
+    protected void deleteInvalidUser(final RealmModel realm, final UserModel user) {
+        String userId = user.getId();
+        String userName = user.getUsername();
         // This needs to be running in separate transaction because removing the user may end up with throwing
         // PessimisticLockException which also rollbacks Jpa transaction, hence when it is running in current transaction
         // it will become not usable for all consequent jpa calls. It will end up with Transaction is in rolled back
@@ -233,7 +239,6 @@ public class UserStorageManager extends AbstractStorageManager<UserStorageProvid
             }
         });
     }
-
 
     protected Stream<UserModel> importValidation(RealmModel realm, Stream<UserModel> users) {
         return users.map(user -> importValidation(realm, user)).filter(Objects::nonNull);
